requirements.txt
amqp==5.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.1
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.1.0
Django==4.2.16
django-cors-headers==4.1.0
django-filter==24.3
django-import-export==3.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.10
jwcrypto==1.5.6
kombu==5.5.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.5
packaging==25.0
pillow==10.4.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-environ==0.4.54
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.1
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
